+++ ErsatzTV.Application/FFmpegProfiles/Commands/CreateFFmpegProfile.cs
@@ -8,6 +8,7 @@ public record CreateFFmpegProfile(
     string Name,
     int ThreadCount,
     HardwareAccelerationKind HardwareAcceleration,
+    string VaapiDisplay,
     VaapiDriver VaapiDriver,
     string VaapiDevice,
     int? QsvExtraHardwareFrames,